+++ src/flowchem/components/analytics/dad.py
