+++ src/flowchem/components/analytics/hplc.py
