+++ src/flowchem/components/analytics/ir.py
