+++ src/flowchem/components/analytics/nmr.py
