+++ src/flowchem/components/sensors/base_sensor.py
