+++ src/flowchem/components/technical/power.py
