+++ src/flowchem/devices/harvardapparatus/elite11_pump.py
@@ -7,7 +7,7 @@ from loguru import logger
 
 if TYPE_CHECKING:
     from .elite11 import Elite11
-from flowchem.components.pumps.syringe_pump import SyringePump
+from flowchem.components.pumps.syringe import SyringePump
 
 
 class Elite11PumpOnly(SyringePump):
