+++ src/flowchem/devices/knauer/azura_compact_pump.py
@@ -9,7 +9,7 @@ from ... import ureg
 
 if TYPE_CHECKING:
     from .azura_compact import AzuraCompact
-from flowchem.components.pumps.hplc_pump import HPLCPump
+from flowchem.components.pumps.hplc import HPLCPump
 
 
 def isfloat(rate: str) -> bool:
