+++ src/flowchem/devices/knauer/dad.py
@@ -3,7 +3,7 @@ import asyncio
 
 from loguru import logger
 
-from flowchem.components.analytics.dad_control import DADControl
+from flowchem.components.analytics.dad import DADControl
 from flowchem.devices.flowchem_device import FlowchemDevice
 from flowchem.devices.knauer._common import KnauerEthernetDevice
 from flowchem.devices.list_known_device_type import autodiscover_third_party
