+++ src/flowchem/devices/mettlertoledo/icir.py
@@ -8,7 +8,7 @@ from asyncua import ua
 from loguru import logger
 from pydantic import BaseModel
 
-from flowchem.components.analytics.ir_control import IRSpectrum
+from flowchem.components.analytics.ir import IRSpectrum
 from flowchem.devices.flowchem_device import DeviceInfo
 from flowchem.devices.flowchem_device import FlowchemDevice
 from flowchem.devices.mettlertoledo.icir_control import IcIRControl
