+++ src/flowchem/devices/mettlertoledo/icir_control.py
@@ -2,8 +2,8 @@ from __future__ import annotations
 
 from typing import TYPE_CHECKING
 
-from flowchem.components.analytics.ir_control import IRControl
-from flowchem.components.analytics.ir_control import IRSpectrum
+from flowchem.components.analytics.ir import IRControl
+from flowchem.components.analytics.ir import IRSpectrum
 
 if TYPE_CHECKING:
     from .icir import IcIR
